Google API modified
Google Historical values is the same as yahoo and must be called from parent.
Google Real time quote has been done, but this one provides only the current or the last day the market opened thus it may not be real time and based on historical values, so this must be checked. If this can't provide the real time quote then this function will call yahoo or tdameritrade.
Main / Test form is now changed to reflect the Google historical values and "real-time" quote.