Creating Skeleton for Bases
Decided to create some skeleton for the bases such as a cup with handle. These bases are dynamically instantiated upon run time and auto-discovered by the base class. Still some more work to be done for the skeleton, but compiles good and finds the derived classes from the base class without issues. This tells us that we can add as many bases as we please with the addition of classes and a proper structure.