TDAmeritrade API continuing integration
Continuing the integration of the API. Currently I am getting very close, but there is no solution. I get the 400 error which the sent request is not authenticating because of either the header or the message not completed. Will have to send an email to TD Ameritrade with this issue.Meanwhile I need to shorten up the program to send the whole solution to TDAmeritrade.